requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
pydantic==2.5.0
# sentence-transformers==3.1.1  # DESACTIVADO: Modelo pesado de embeddings
python-multipart==0.0.6
python-dotenv==1.0.0
# numpy==1.24.3  # DESACTIVADO: Solo necesario para embeddings
# google-generativeai==0.3.2  # DESACTIVADO: API de Gemini